requirements.txt
Django==1.7c2
Jinja2==2.7.3
MarkupSafe==0.23
Pygments==1.6
Sphinx==1.2.2
django-extensions==1.3.8
django-shell-plus==1.1.5
djangorestframework==2.3.14
docutils==0.12
mock==1.0.1
model-mommy==1.2.1
six==1.3.0
wsgiref==0.1.2
